yggdrasil/yggdrasil/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.4
distlib==0.3.8
Django==4.2.7
django-extensions==3.2.3
django-filter==23.4
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
pipenv==2023.11.15
platformdirs==4.1.0
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
shellingham==1.4.0
sqlparse==0.4.4
toml==0.10.2
typer==0.3.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
markdown-it-py==3.0.0
mdurl==0.1.2
rich==13.7.0
Pygments==2.17.2